Extract instance::settings and instance::settings::effective to a concern
